[moviebrowser] -show title as info2  if this is missed

+++ branches/nhd2-exp/lib/libdvbapi/playback_cs.h
@@ -49,7 +49,7 @@ class cPlayback
 		bool SetSlow(int slow);
 		bool GetSpeed(int &speed) const;
 		bool GetPosition(int64_t &position, int64_t &duration);
-		bool SetPosition(int position);
+		bool SetPosition(int64_t position);
 		void FindAllPids(uint16_t *apids, unsigned short *ac3flags, uint16_t *numpida, std::string *language);
 
 		cPlayback(int num = 0);
